refactor(#1371): extract startup_local_runtime inline tests
